+++ plugins/main/public/controllers/management/components/management/configuration/registration-service/registration-service.js
@@ -39,32 +39,34 @@ const mainSettings = [
   { field: 'port', label: 'Listen to connections at port' },
   { field: 'use_source_ip', label: "Use client's source IP address" },
   { field: 'use_password', label: 'Use a password to register agents' },
-  { field: 'purge', label: 'Purge agents list when removing agents' },
+  { field: 'purge', label: 'Purge agents list on removal' },
   {
     field: 'limit_maxagents',
-    label: 'Limit registration to maximum number of agents',
+    label: 'Limit registration to max agents',
   },
   {
     field: 'force.enabled',
-    label: 'Force registration when using an existing IP address',
+    label: 'Force registration on existing IP',
   },
   {
     field: 'force.after_registration_time',
-    label:
-      'Specifies that the agent replacement will be performed only when the time (seconds) passed since the agent registration is greater than the value configured in the setting',
+    label: 'Min seconds since registration',
+    info: 'The agent replacement is performed only when the time (in seconds) passed since the agent registration is greater than the value configured in the setting.',
   },
   {
     field: 'force.key_mismatch',
-    label: 'Avoid re-registering agents that already have valid keys',
+    label: 'Re-register only on key mismatch',
+    info: 'Avoid re-registering agents that already have valid keys.',
   },
   {
     field: 'force.disconnected_time.enabled',
-    label:
-      'Specifies that the replacement will be performed only for agents that have been disconnected longer than a certain time',
+    label: 'Replace only disconnected agents',
+    info: 'The replacement is performed only for agents that have been disconnected longer than a certain time.',
   },
   {
     field: 'force.disconnected_time.value',
-    label: 'Seconds since an agent is in a disconnected state',
+    label: 'Seconds an agent is disconnected',
+    info: 'Seconds since an agent is in a disconnected state.',
   },
 ];